Add Youtube Video Publication Feature
Detection of Youtube link (with regex) in post publication form, if detected Iframe preview appear. 
Creation of new column "video_url" in DB and modification of route, controller and model on backend side. 
Add detection of links in general (with regex)  in post content. If link detected <a> tag with link href is created and link is cut from post content.